Rename package name to themeetgroup
