+++ src/main/java/com/github/themeetgroup/kafka/connect/rabbitmq/CommonRabbitMQConnectorConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jcustenborder.kafka.connect.rabbitmq;
+package com.github.themeetgroup.kafka.connect.rabbitmq;
 
 import com.rabbitmq.client.ConnectionFactory;
 import org.apache.kafka.common.config.AbstractConfig;
