+++ src/main/java/com/github/themeetgroup/kafka/connect/rabbitmq/source/data/SourceMessageConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jcustenborder.kafka.connect.rabbitmq.source.data;
+package com.github.themeetgroup.kafka.connect.rabbitmq.source.data;
 
 import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Envelope;
