+++ src/main/java/com/github/themeetgroup/kafka/connect/rabbitmq/source/data/SourceRecordBuilder.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jcustenborder.kafka.connect.rabbitmq.source.data;
+package com.github.themeetgroup.kafka.connect.rabbitmq.source.data;
 
-import com.github.jcustenborder.kafka.connect.rabbitmq.source.RabbitMQSourceConnectorConfig;
+import com.github.themeetgroup.kafka.connect.rabbitmq.source.RabbitMQSourceConnectorConfig;
 import com.google.common.collect.ImmutableMap;
 import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.Envelope;
