+++ src/test/java/com/github/themeetgroup/kafka/connect/rabbitmq/source/data/MessageConverterTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.jcustenborder.kafka.connect.rabbitmq.source.data;
+package com.github.themeetgroup.kafka.connect.rabbitmq.source.data;
 
 import com.google.common.collect.ImmutableMap;
 import com.rabbitmq.client.BasicProperties;
