+++ src/test/resources/logback.xml
@@ -7,5 +7,5 @@
     <root level="INFO">
         <appender-ref ref="STDOUT" />
     </root>
-    <logger name="com.github.jcustenborder.kafka.connect.rabbitmq" level="TRACE" />
+    <logger name="com.github.themeetgroup.kafka.connect.rabbitmq" level="TRACE" />
 </configuration>
